+++ MSS_DEMO/MSS.csproj
@@ -183,6 +183,7 @@
     <Compile Include="App_Start\BundleConfig.cs" />
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
+    <Compile Include="Controllers\Report\ReportController.cs" />
     <Compile Include="Controllers\UserManagement\User_RoleController.cs" />
     <Compile Include="Core\Implement\ClassStudentRepository.cs" />
     <Compile Include="Core\Implement\CoursesDeadlineRepository.cs" />
@@ -194,7 +195,6 @@
     <Compile Include="IoC\UnityDI.cs" />
     <Compile Include="Controllers\Log\Student_Course_LogController.cs" />
     <Compile Include="Controllers\Log\Student_Specification_LogController.cs" />
-    <Compile Include="Controllers\ReportController.cs" />
     <Compile Include="Controllers\SetUp\CampusController.cs" />
     <Compile Include="Controllers\SetUp\CoursesController.cs" />
     <Compile Include="Controllers\Import\ImportDataDailyController.cs" />
@@ -387,6 +387,8 @@
     <Content Include="Models\MSSModel.edmx.diagram">
       <DependentUpon>MSSModel.edmx</DependentUpon>
     </Content>
+    <Content Include="Views\Report\Enrollment.cshtml" />
+    <Content Include="Views\Report\Member.cshtml" />
   </ItemGroup>
   <ItemGroup />
   <ItemGroup>
